+++ versioned_hdf5/replay.py
@@ -204,13 +204,23 @@ def _recreate_raw_data(
     raw_data = f['_version_data'][name]['raw_data']
     chunks = ChunkSize(raw_data.chunks)
     new_shape = (len(chunks_to_keep)*chunks[0], *chunks[1:])
+    dtype = raw_data.dtype
 
+    fillvalue = raw_data.fillvalue
+    if dtype.metadata and ('vlen' in dtype.metadata or 'h5py_encoding' in dtype.metadata):
+        # h5py string dtype
+        # (https://h5py.readthedocs.io/en/2.10.0/strings.html). Setting the
+        # fillvalue in this case doesn't work
+        # (https://github.com/h5py/h5py/issues/941).
+        if fillvalue not in [0, '', b'', None]:
+            raise ValueError("Non-default fillvalue not supported for variable length strings")
+        fillvalue = None
     new_raw_data = f['_version_data'][name].create_dataset(
         '_tmp_raw_data', shape=new_shape, maxshape=(None,)+chunks[1:],
-        chunks=raw_data.chunks, dtype=raw_data.dtype,
+        chunks=raw_data.chunks, dtype=dtype,
         compression=raw_data.compression,
         compression_opts=raw_data.compression_opts,
-        fillvalue=raw_data.fillvalue)
+        fillvalue=fillvalue)
     for key, val in raw_data.attrs.items():
         new_raw_data.attrs[key] = val