+++ versioned_hdf5/tests/test_replay.py
@@ -3,7 +3,12 @@ import sys
 
 import h5py
 import numpy as np
+import pathlib
+import subprocess
+import shutil
+import pytest
 
+from versioned_hdf5        import VersionedHDF5File
 from versioned_hdf5.replay import (modify_metadata, delete_version,
                                    delete_versions, _recreate_raw_data,
                                    _recreate_hashtable,
@@ -826,6 +831,50 @@ def test_delete_empty_dataset(vfile):
 
     assert vfile.f['_version_data/key0/raw_data'][:].size == 10000
     assert vfile[vfile.current_version]['key0'][:].size == 0
+
+@pytest.mark.skipif(shutil.which('h5repack') is None, reason="Requires h5repack to run")
+def test_delete_string_dataset(filepath):
+    """Test that delete_versions + h5repack works correctly for variable length string
+    dtypes.
+
+    When calling delete_versions, the dataset must be reconstructed from the remaining
+    versions using a NoneType fillvalue. However, because we can't store a NoneType for
+    the fillvalue of the dataset in the h5 file, it is instead stored as b''. Previously
+    a bug in delete_versions would recreate the datset using the file's fillvalue of
+    b'' rather than None, corrupting the data. See https://github.com/h5py/h5py/issues/941
+    for more information about the bug in h5py responsible for this, and
+    https://github.com/deshaw/versioned-hdf5/issues/238 for the versioned-hdf5
+    discussion.
+    """
+    # Create two versions of variable-length string data, then delete the first,
+    # forcing a reconstruction of the data
+    with h5py.File(filepath, 'w') as f:
+        vf = VersionedHDF5File(f)
+        with vf.stage_version('r0') as sv:
+            sv.create_dataset(
+                'foo',
+                data=['abc'],
+                chunks=(100,),
+                dtype=h5py.string_dtype(encoding='ascii'),
+            )
+
+        with vf.stage_version('r1') as sv:
+            sv['foo'][0] = 'def'
+
+        delete_versions(f, 'r0')
+
+    # Repack the data; the RuntimeError only appears if you do this
+    tmp_path = pathlib.Path(filepath + '.tmp')
+    subprocess.run(['h5repack', filepath, tmp_path], stdout=subprocess.PIPE)
+
+    # Check that staging a new version after delete_versions + h5repack works
+    with h5py.File(tmp_path, 'r+') as f:
+        vf = VersionedHDF5File(f)
+        with vf.stage_version('r3') as sv:
+            pass
+
+    # Delete the tmp file to avoid leaving around stale file
+    tmp_path.unlink(missing_ok=True)
 
 
 def test_delete_versions_speed(vfile):